% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aez}
\alias{aez}
\title{GRDC Australian Grains Agroecological Zones}
\format{
An \code{\link[sf:sf]{sf::sf()}} polygon object of the \acronym{GRDC} Australian Grains
Agroecological Zones with 18 rows and two variables.
\describe{
\item{AEZ}{The agroecological zone name.}
\item{geometry}{The centroid of the agroecological zone.}
}
}
\source{
The Australian Grains Research and Development Corporation,
\acronym{GRDC}, \url{https://grdc.com.au/}, under CC0 licence.
}
\usage{
aez
}
\description{
GRDC Australian Grains Agroecological Zones
}
\section{Coordinate Reference System}{

Geodetic CRS:  WGS 84
}

\keyword{datasets}
