% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_area.R
\name{extract_area}
\alias{extract_area}
\title{Extract an Area of Interest Using Australian GPS Coordinates}
\usage{
extract_area(x, coords, spatial, area)
}
\arguments{
\item{x}{a \code{data.frame} of \acronym{GPS} coordinates in two columns.}

\item{coords}{names or numbers of the numeric columns holding coordinates in
\code{lonlat}.}

\item{spatial}{a user-supplied \link{sf} object that contains information to
derive location information from.}

\item{area}{the field in \code{spatial} that should be returned.}
}
\value{
a \code{data.frame} with the provided \acronym{GPS} coordinates and the
respective \code{area} value from \code{spatial}.
}
\description{
A generic function to extract an area of interest for user-provided
\acronym{GPS} coordinates.
}
\examples{
# load the `aez` data included in the package for use in example only.
# the `extract_ae_zone()` performs this exact task, this is strictly
# for demonstration purposes only
library(sf)
aez <- read_sf(system.file(
"extdata",
"aez.gpkg",
package = "extractOz",
mustWork = TRUE
))
locs <- data.frame(
  site = c("Merredin", "Corrigin", "Tamworth"),
  "x" = c(118.28, 117.87, 150.84),
  "y" = c(-31.48, -32.33, -31.07)
)
extract_area(x = locs, coords = c("x", "y"), spatial = aez, area = "AEZ")
}
\seealso{
Other extract functions: 
\code{\link{extract_ae_zone}()},
\code{\link{extract_soil_order}()}
}
\concept{extract functions}
