% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_smips.R
\name{extract_smips}
\alias{extract_smips}
\title{Extract Soil Moisture Data from TERN's SMIPS Using Australian GPS Coordinates}
\usage{
extract_smips(
  x,
  data = "smips",
  collection = "totalbucket",
  day,
  api_key = nert::get_key(),
  max_tries = 3,
  initial_delay = 1
)
}
\arguments{
\item{x}{\code{List}. An object with names containing \code{vector} pairings of
longitude and latitude values expressed as decimal degree values in that
order, each individual \code{vectors}' items should be named "x" (longitude) and
"y" (latitude), respectively.  The \code{list} item names should be descriptive
of the individual \code{vectors} and will be included in a "location" column of
the output.}

\item{data}{A character vector of the data source to be queried, currently
only \dQuote{smips}.}

\item{collection}{A character vector of the data collection to be queried,
currenly only \dQuote{smips} is supported with the following collections:
\itemize{
\item SMindex
\item bucket1
\item bucket2
\item deepD
\item runoff
\item totalbucket
}

Defaults to \dQuote{totalbucket}. Multiple \code{collections} are supported,
\emph{e.g.}, \code{collection = c("SMindex", "totalbucket")}.}

\item{day}{A vector of date(s) to query, \emph{e.g.}, \code{day = "2017-12-31"} or
\code{day = seq.Date(as.Date("2017-12-01"), as.Date("2017-12-31"), "days")}, both
\code{Character} and \code{Date} classes are accepted.}

\item{api_key}{A \code{character} string containing your \acronym{API} key,
a random string provided to you by \acronym{TERN}, for the request.
Defaults to automatically detecting your key from your local .Renviron,
.Rprofile or similar.  Alternatively, you may directly provide your key as
a string here or use functionality like that from \CRANpkg{keyring}.  If
nothing is provided, you will be prompted on how to set up your \R session
so that it is auto-detected and a browser window will open at the
\acronym{TERN} website for you to request a key.}

\item{max_tries}{An integer \code{Integer} with the number of times to retry a
failed download before emitting an error message.  Defaults to 3.}

\item{initial_delay}{An \code{Integer} with the number of seconds to delay before
retrying the download.  This increases as the tries increment.  Defaults to
1.}
}
\value{
A \link[data.table:data.table]{data.table::data.table} with the provided \acronym{GPS}
coordinates and the respective soil moisture data from \acronym{SMIPS} and
coordinates from the \acronym{SMIPS} ("smips_longitude" and
"smips_latitude")
}
\description{
Extracts the soil moisture at the GPS points provided assuming that they are
land-based coordinates in Australia from the Soil Moisture Integration and
Prediction System \acronym{SMIPS} data set.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
locs <- list(
  "Merredin" = c(x = 118.28, y = -31.48),
  "Corrigin" = c(x = 117.87, y = -32.33),
  "Tamworth" = c(x = 150.84, y = -31.07)
)

extract_smips(x = locs, day = "2023-09-23")
\dontshow{\}) # examplesIf}
}
\references{
\url{https://portal.tern.org.au/metadata/TERN/15728dba-b49c-4da5-9073-13d8abe67d7chttps://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0}
}
