% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_order.R
\name{extract_soil_order}
\alias{extract_soil_order}
\title{Extract Soil Order from DAAS using Australian GPS Coordinates}
\usage{
extract_soil_order(x, coords)
}
\arguments{
\item{x}{a \code{data.frame} of \acronym{GPS} coordinates in two columns.}

\item{coords}{names or numbers of the numeric columns holding coordinates in
\code{lonlat}.}
}
\value{
a \code{data.frame} with the provided \acronym{GPS} coordinates and the
respective Digital Atlas of Australian Soils (\acronym{DAAS} soil order).
}
\description{
Extract Soil Order from DAAS using Australian GPS Coordinates
}
\examples{
locs <- data.frame(
  site = c("Merredin", "Corrigin", "Tamworth"),
  "x" = c(118.28, 117.87, 150.84),
  "y" = c(-31.48, -32.33, -31.07)
)
extract_soil_order(x = locs, coords = c("x", "y"))
}
\seealso{
Other extract functions: 
\code{\link{extract_ae_zone}()}
}
\concept{extract functions}
