% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weather.R
\name{extract_weather}
\alias{extract_weather}
\title{Extract Weather from SILO and DPIRD Weather Station Networks from Australian GPS Coordinates}
\usage{
extract_weather(
  x,
  first,
  last,
  which_api = "silo",
  silo_api_key = NULL,
  dpird_api_key = NULL
)
}
\arguments{
\item{x}{\code{List}. An object with names containing \code{vector} pairings of
longitude and latitude values expressed as decimal degree values in that
order, each individual \code{vectors}' items should be named "x" (longitude) and
"y" (latitude), respectively.  The \code{list} item names should be descriptive
of the individual \code{vectors} and will be included in a "location" column of
the output.}

\item{first}{\code{Integer}. A value representing the start date of the query in
the format 'yyyymmdd' (ISO-8601).}

\item{last}{\code{Integer}. A value representing the end date of the query in the
format 'yyyymmdd' (ISO-8601).}

\item{which_api}{\code{Character}. A \code{string} value that indicates which API to
use.  Defaults to "silo".  Valid values are "all", for both \acronym{SILO}
(\acronym{BOM}) and \acronym{DPIRD} weather station networks; "silo" for
only stations in the \acronym{SILO} network; or "dpird" for stations in the
\acronym{DPIRD} network.}

\item{silo_api_key}{\code{Character}. A string specifying a valid email address to
use for the request. The query will return an error if a valid email address
is not provided.  Only used when \code{which_api} is set to "silo" or to "all".}

\item{dpird_api_key}{\code{Character}. A \code{string} value that is the user's
\acronym{API} key from \acronym{DPIRD} (see
\url{https://www.agric.wa.gov.au/web-apis}).  Only used when \code{which_api} is
"dpird" or "all".}

\item{interval}{\code{Character}. A \code{string} value that indicates whether to
return daily or monthly interval data.  Valid values are "daily" or
"monthly".  Defaults to "daily".}
}
\value{
A \code{data.table} of weather station data for the weather stations
nearest the provided longitude and latitude values for the requested weather
station network(s) with the following columns:
\tabular{rl}{
\strong{location}:\tab User supplied name for GPS coordinates in \code{x}.\cr
\strong{station_number}:\tab Station ID as in the respective network.\cr
\strong{station_name}:\tab Station name in the respective network.\cr
\strong{date}:\tab The observation date as an integer, "yyyymmdd".\cr
\strong{year}:\tab Year of observation (YYYY).\cr
\strong{month}:\tab Month of observation (1-12).\cr
\strong{day}:\tab Day of month of observation (1-31).\cr
\strong{jday}:\tab Day of year of the observation. Commonly referred to as the
Julian date.\cr
\strong{min_temperature}:\tab Minimum daily recorded temperature (degrees C).\cr
\strong{max_temperature}:\tab Maximum daily recorded temperature (degrees C).\cr
\strong{accum_days_min}:\tab Accumulated number of days of minimum
temperature.\cr
\strong{accum_days_max}:\tab Accumulated number of days of maximum
temperature.\cr
\strong{rainfall}:\tab Daily recorded rainfall in mm.\cr
\strong{period}:\tab Period over which rainfall was measured.\cr
\strong{solar_exposure}:\tab Daily global solar exposure in MJ/m^2.\cr
\strong{longitude}:\tab Station longitude in decimal degrees (WGS84).\cr
\strong{latitude}:\tab Station latitude in decimal degrees (WGS84).\cr
\strong{state}:\tab State that the station is located in.\cr
\strong{elevation}:\tab Station elevation in metres.\cr
\strong{owner}:\tab Station owner.\cr
\strong{distance}:\tab Distance from specified geolocation in \code{x}.\cr
}
}
\description{
Extracts weather data for the provided dates from the nearest weather station
to the provided longitude and latitude values in \code{x}.  Two \acronym{APIs}
are available, "SILO" and "DPIRD".  Either one or both may be used in the
query.  Weather data for the single closest station to the provided
longitude and latitude will always be returned from the selected
\acronym{API(s)}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Source data from a list of latitude and longitude coordinates in NSW and WA
locs <- list(
  "Merredin" = c(x = 118.28, y = -31.48),
  "Corrigin" = c(x = 117.87, -32.33),
  "Tamworth" = c(x = 150.84, y = -31.07)
)

# Replace "YOUR API KEY" with a valid DPIRD API key and "YOUR EMAIL"
# with a valid email address.

w <- extract_weather(
  x = locs,
  first = "20210101",
  last = "20210131",
  interval = "daily",
  which_api = "all",
  dpird_api_key = "YOUR API KEY",
  silo_api_key = "YOUR EMAIL"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other weather data: 
\code{\link{extract_power}()},
\code{\link{extract_silo}()}
}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
\concept{weather data}
