% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOz-package.R
\docType{package}
\name{extractOz-package}
\alias{extractOz}
\alias{extractOz-package}
\title{extractOz: Extract Australian Soil, Weather and Agro-ecological Zone Information from GPS Points}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Extracts point data for a given GPS coordinate that includes major soil order, weather and GRDC agro-ecological zone information at that point.
}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adam.sparks@dpird.wa.gov.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Other contributors:
\itemize{
  \item Western Australia Department of Primary Industries and Regional Development [copyright holder]
}

}
\keyword{internal}
