% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_silo_data.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_silo}
\alias{get_dpird_summaries}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{weatherOz}{\code{\link[weatherOz]{get_dpird_summaries}}, \code{\link[weatherOz]{get_silo}}}
}}

