% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aez.R
\docType{data}
\name{aez}
\alias{aez}
\title{GRDC Australian Grains Agroecological Zones}
\format{
An \code{\link[sf:sf]{sf::sf()}} polygon object
}
\source{
The Australian Grains Research and Development Corporation,
\acronym{GRDC}, \url{https://grdc.com.au/} under CC0 licence.
}
\usage{
aez
}
\description{
An \link[sf:sf]{sf::sf} object for displaying GRDC Australian Grains Agroecological
Zones
}
\section{Coordinate Reference System}{

Geodetic CRS:  WGS 84
}

\keyword{datasets}
