% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_list.R
\name{df_to_list}
\alias{df_to_list}
\title{Convert a data.frame or Tibble to a list suitable for use in extractOz}
\usage{
df_to_list(x, names = 1, lonlat = 2:3)
}
\arguments{
\item{x}{\code{data.frame}. A three column object with the first column being the
location name, the second longitude, named "x", and the third, latitude,
named "y", as decimal degrees in \code{numeric} format.}

\item{names}{\code{Integer}. The column in which the location names are located.
Defaults to 1.}

\item{lonlat}{\code{Vector}. A set of \code{integer} values indicating the column index
in which the longitude and latitude values are contained. Defaults to \code{2:3}.}
}
\value{
A \code{list} object suitable for use in any of the \code{extract} functions
in \pkg{extractOz}.
}
\description{
A helper function to convert \code{data.frames} or objects inheriting from
\code{data.frame}, \emph{e.g.} \code{\link[tibble:tibble]{tibble::tibble()}} or \code{\link[data.table:data.table]{data.table::data.table()}} into
a list suitable for use in \pkg{extractOz} functions.
}
\examples{
library(readr)

read_csv(system.file(
  "extdata",
  "sample_points.csv",
  package = "extractOz",
  mustWork = TRUE
)) |>
  df_to_list(names = 1, indices = 2:3)

}
\author{
Adam H. Sparks, \email{adam.sparks@dpird.wa.gov.au}
}
