% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ae_zone.R
\name{extract_ae_zone}
\alias{extract_ae_zone}
\title{Extract GRDC Agroecological Zones Using Australian GPS Coordinates}
\usage{
extract_ae_zone(x)
}
\arguments{
\item{x}{\code{List}. An object with names containing \code{vector} pairings of
longitude and latitude values expressed as decimal degree values in that
order, each individual \code{vectors}' items should be named "x" (longitude) and
"y" (latitude), respectively.  The \code{list} item names should be descriptive
of the individual \code{vectors} and will be included in a "location" column of
the output.}
}
\value{
A \code{data.table} with the provided \acronym{GPS} coordinates and the
respective \acronym{GRDC} agroecological zone.
}
\description{
Extract GRDC Agroecological Zones Using Australian GPS Coordinates
}
\examples{

locs <- list(
  "Merredin" = c(x = 118.28, y = -31.48),
  "Corrigin" = c(x = 117.87, y = -32.33),
  "Tamworth" = c(x = 150.84, y = -31.07)
  )

extract_ae_zone(x = locs)

}
