% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ae_zone.R
\name{extract_ae_zone}
\alias{extract_ae_zone}
\title{Extract GRDC Agroecological Zones Using Australian GPS Coordinates}
\usage{
extract_ae_zone(x, coords)
}
\arguments{
\item{x}{a \code{data.frame} of \acronym{GPS} coordinates in two columns.}

\item{coords}{names or numbers of the numeric columns holding coordinates in
\code{lonlat}.}
}
\value{
a \code{data.frame} with the provided \acronym{GPS} coordinates and the
respective \acronym{GRDC} agroecological zone.
}
\description{
Extract GRDC Agroecological Zones Using Australian GPS Coordinates
}
\examples{
locs <- data.frame(
  site = c("Merredin", "Corrigin", "Tamworth"),
  "x" = c(118.28, 117.87, 150.84),
  "y" = c(-31.48, -32.33, -31.07)
)
extract_ae_zone(x = locs, coords = c("x", "y"))
}
\seealso{
Other extract functions: 
\code{\link{extract_area}()},
\code{\link{extract_soil_order}()}
}
\concept{extract functions}
