% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_silo.R
\name{extract_silo}
\alias{extract_silo}
\title{Extract Weather from SILO PatchedPointData using Australian GPS Coordinates}
\usage{
extract_silo(x, first, last, data_format = "alldata", email)
}
\arguments{
\item{x}{\code{List}. An object with names containing \code{vector} pairings of
longitude and latitude values expressed as decimal degree values in that
order, each individual \code{vectors}' items should be named "x" (longitude) and
"y" (latitude), respectively.  The \code{list} item names should be descriptive
of the individual \code{vectors} and will be included in a "location" column of
the output.}

\item{first}{\code{Integer}. A string representing the start date of the query in
the format 'yyyymmdd' (ISO-8601).}

\item{last}{\code{Integer}. A string representing the end date of the query in the
format 'yyyymmdd' (ISO-8601).}

\item{data_format}{\code{Character}. A string specifying the type of data to
retrieve.  Valid values are 'alldata', 'monthly' or 'apsim' (formatted for
use in APSIM modelling).  Note that 'apsim' and 'alldata' only retrieve
daily time-step data.}

\item{email}{\code{Character}. A string specifying a valid email address to use
for the request.  The query will return an error if a valid email address is
not provided.}
}
\value{
A \code{data.table} containing the retrieved data from the \acronym{SILO}
\acronym{API}.
}
\description{
A modified wrapper version of \code{\link[weatherOz:get_silo]{weatherOz::get_silo()}} that allows for fetching
many geophysical points or a single geophysical point.  Extracts weather data
from the \acronym{SILO} \acronym{API} from the gridded data
(PatchedPointData).  There are three formats available: 'alldata' and 'apsim'
with daily frequency and 'monthly' with, that's right, monthly frequency.
}
\note{
Where multiple \code{lonlat} values are provided in a single request and the
\code{data_format} is \code{apsim}, an extra column, "location" will be provided with
the list names names as the values.  Users may then use \code{split()} to create a
list of \code{data.table} objects that drops this column or \code{subset()} and remove
the "location" column and enable the use of the data in APSIM modelling work.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Source data from a list of latitude and longitude coordinates in NSW and WA
locs <- list(
  "Merredin" = c(x = 118.28, y = -31.48),
  "Corrigin" = c(x = 117.87, -32.33),
  "Tamworth" = c(x = 150.84, y = -31.07)
)

wd <- extract_silo(x = locs,
                    first = "20211001",
                    last = "20211201",
                    data_format = "apsim",
                    email = "YOUR EMAIL")
\dontshow{\}) # examplesIf}
}
