% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_order.R
\name{extract_soil_order}
\alias{extract_soil_order}
\title{Extract Soil Order from DAAS using Australian GPS Coordinates}
\source{
\url{https://data.gov.au/dataset/ds-dga-5ccb44bf-93f2-4f94-8ae2-4c3f699ea4e7/distribution/dist-dga-56ba5f25-2324-43b5-8df8-b9c69ae2ea0b/details?q=}
}
\usage{
extract_soil_order(x, coords)
}
\arguments{
\item{x}{a \code{data.frame} of \acronym{GPS} coordinates in two columns.}

\item{coords}{names or numbers of the numeric columns holding coordinates in
\code{lonlat}.

Extracts the major soil order at the GPS points provided.}
}
\value{
a \code{data.frame} with the provided \acronym{GPS} coordinates and the
respective Digital Atlas of Australian Soils (\acronym{DAAS} soil order),
"Spatial Data Conversion of the Atlas of Australian Soils to the Australian
Soil Classification v01".
}
\description{
Extract Soil Order from DAAS using Australian GPS Coordinates
}
\note{
The first run will take additional time to download and extract the
soils data and cache it locally for future use. Any use after this will be
much faster due to the locally cached geospatial soils data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
locs <- data.frame(
  site = c("Merredin", "Corrigin", "Tamworth"),
  "x" = c(118.28, 117.87, 150.84),
  "y" = c(-31.48, -32.33, -31.07)
)
extract_soil_order(x = locs, coords = c("x", "y"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other extract functions: 
\code{\link{extract_ae_zone}()},
\code{\link{extract_area}()}
}
\concept{extract functions}
