% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOz-package.R
\docType{package}
\name{extractOz-package}
\alias{extractOz}
\alias{extractOz-package}
\title{extractOz: Extract Australian Soil, Weather and GRDC Agro-ecological Zone Information from GPS Points}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extracts point data for a given GPS coordinate that includes major soil order, weather and GRDC agro-ecological zone information at that point.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DPIRD-FSI/extractOz}
  \item Report bugs at \url{https://github.com/DPIRD-FSI/extractOz/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adam.sparks@dpird.wa.gov.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Other contributors:
\itemize{
  \item Western Australia Agriculture Authority (WAAA) (Provided support through Adam Sparks's time.) [copyright holder]
  \item Grains Research and Development Corporation (GRDC Project GRI2007-001RTX) [funder, copyright holder]
}

}
\keyword{internal}
