% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_silo_data.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_silodata}
\alias{get_multi_silodata}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cropgrowdays}{\code{\link[cropgrowdays]{get_multi_silodata}}, \code{\link[cropgrowdays]{get_silodata}}}
}}

