% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_order.R
\name{remove_cache}
\alias{remove_cache}
\title{Remove DAAS Soil Order Cache}
\usage{
remove_cache()
}
\value{
\code{NULL}, called for its side-effects
}
\description{
The cache is located at \code{tools::R_user_dir("extractOz", which = "cache")}
}
